Cập nhật ProductService
